+++ python/common/models/TAR/additional_collision_details.py
