+++ python/common/models/TAR/collision.py
