+++ python/common/models/TAR/collision_location.py
