+++ python/common/models/TAR/damage_location.py
