+++ python/common/models/TAR/lighting_condition.py
