+++ python/common/models/TAR/safety_equipment.py
