+++ python/common/models/TAR/weather_condition.py
