+++ python/common/models/__init__.py
@@ -29,43 +29,43 @@ from .impound_reason_codes import ImpoundReasonCodes
 from .ilo_id_crossref import IloIdCrossRef
 from .df_errors import DFErrors
 from .submission import Submission
-from .TAR.tar_collision import TarCollision
-from .TAR.tar_additional_collision_details import TarAdditionalCollisionDetails
-from .TAR.tar_location import TarLocation
-from .TAR.tar_witness_info import TarWitnessInfo
-from .TAR.tar_entity import TarEntity
-from .TAR.tar_entity_vehicle import TarEntityVehicle
-from .TAR.tar_involved_person import TarInvolvedPerson
-from .TAR.tar_charges import TarCharges
-from .TAR.tar_road_class import TarRoadClass
-from .TAR.tar_traffic_flow import TarTrafficFlow
-from .TAR.tar_collision_location import TarCollisionLocation
-from .TAR.tar_speed_zone import TarSpeedZone
-from .TAR.tar_land_usage import TarLandUsage
-from .TAR.tar_road_type import TarRoadType
-from .TAR.tar_traffic_control import TarTrafficControl
-from .TAR.tar_roadway_character import TarRoadwayCharacter
-from .TAR.tar_roadway_condition import TarRoadwayCondition
-from .TAR.tar_weather_condition import TarWeatherCondition
-from .TAR.tar_lighting_condition import TarLightingCondition
-from .TAR.tar_type_of_collision import TarTypeOfCollision
-from .TAR.tar_location_of_first_contact import TarLocationOfFirstContact
-from .TAR.tar_primary_collision_occurrence import TarPrimaryCollisionOccurrence
-from .TAR.tar_pedestrian_location import TarPedestrianLocation
-from .TAR.tar_pedestrian_action import TarPedestrianAction
-from .TAR.tar_entity_type import TarEntityType
-from .TAR.tar_contributing_factors import TarContributingFactors
-from .TAR.tar_damage_location import TarDamageLocation
-from .TAR.tar_damage_severity import TarDamageSeverity
-from .TAR.tar_pre_collision_action import TarPreCollisionAction
-from .TAR.tar_vehicle_type import TarVehicleType
-from .TAR.tar_vehicle_use import TarVehicleUse
-from .TAR.tar_position import TarPosition
-from .TAR.tar_safety_equipment import TarSafetyEquipment
-from .TAR.tar_ejection import TarEjection
-from .TAR.tar_injury_location import TarInjuryLocation
-from .TAR.tar_injury_type import TarInjuryType
-from .TAR.tar_victim_status import TarVictimStatus
-from .TAR.tar_taken_to import TarTakenTo
-from .TAR.tar_taken_by import TarTakenBy
-from .TAR.tar_injury_classification import TarInjuryClassification
+from .TAR.collision import TarCollision
+from .TAR.additional_collision_details import TarAdditionalCollisionDetails
+from .TAR.location import TarLocation
+from .TAR.witness_info import TarWitnessInfo
+from .TAR.entity import TarEntity
+from .TAR.entity_vehicle import TarEntityVehicle
+from .TAR.involved_person import TarInvolvedPerson
+from .TAR.charges import TarCharges
+from .TAR.road_class import TarRoadClass
+from .TAR.traffic_flow import TarTrafficFlow
+from .TAR.collision_location import TarCollisionLocation
+from .TAR.speed_zone import TarSpeedZone
+from .TAR.land_usage import TarLandUsage
+from .TAR.road_type import TarRoadType
+from .TAR.traffic_control import TarTrafficControl
+from .TAR.roadway_character import TarRoadwayCharacter
+from .TAR.roadway_condition import TarRoadwayCondition
+from .TAR.weather_condition import TarWeatherCondition
+from .TAR.lighting_condition import TarLightingCondition
+from .TAR.type_of_collision import TarTypeOfCollision
+from .TAR.location_of_first_contact import TarLocationOfFirstContact
+from .TAR.primary_collision_occurrence import TarPrimaryCollisionOccurrence
+from .TAR.pedestrian_location import TarPedestrianLocation
+from .TAR.pedestrian_action import TarPedestrianAction
+from .TAR.entity_type import TarEntityType
+from .TAR.contributing_factors import TarContributingFactors
+from .TAR.damage_location import TarDamageLocation
+from .TAR.damage_severity import TarDamageSeverity
+from .TAR.pre_collision_action import TarPreCollisionAction
+from .TAR.vehicle_type import TarVehicleType
+from .TAR.vehicle_use import TarVehicleUse
+from .TAR.position import TarPosition
+from .TAR.safety_equipment import TarSafetyEquipment
+from .TAR.ejection import TarEjection
+from .TAR.injury_location import TarInjuryLocation
+from .TAR.injury_type import TarInjuryType
+from .TAR.victim_status import TarVictimStatus
+from .TAR.taken_to import TarTakenTo
+from .TAR.taken_by import TarTakenBy
+from .TAR.injury_classification import TarInjuryClassification
